fix(heartbeat): satisfy golangci-lint on files touched by api monitor PR
The api-monitor PR enabled CI's `--whole-files` golangci-lint mode on
the files it modified, surfacing seven pre-existing violations (plus
several more masked by `max-same-issues: 3`) that had been ignored
while those files were untouched on main:

  * heartbeat/monitors/logger/logger.go: `extractRunInfo` was doing
    five unchecked type assertions on `interface{}` values pulled out
    of the event, e.g. `monType.(string)`, which trip `errcheck`'s
    `check-type-assertions: true` and would panic on a malformed
    event. Switched to the `, ok` form and append a typed error to
    the existing aggregated-error path so callers see "monitor.type
    is not a string, got <T>" instead of a runtime panic.

  * heartbeat/monitors/wrappers/summarizer/summarizer.go: changed
    `Summarizer.contsRemaining` from `uint16` to `int` so that
    `contsRemaining += len(conts)` no longer needs the
    `uint16(len(conts))` narrowing conversion that gosec G115 flags.
    The field is only used as an internal "still to process" counter
    compared against 0; widening to int is semantically identical.

  * heartbeat/monitors/factory.go, heartbeat/config/config.go,
    heartbeat/monitors/logger/logger.go,
    x-pack/heartbeat/monitors/browser/sourcejob.go,
    x-pack/heartbeat/monitors/browser/synthexec/synthtypes.go:
    annotated the remaining `logp.L()` fallbacks with targeted
    `//nolint:forbidigo` directives that explain why the call site
    doesn't have a contextual `*logp.Logger` (factory predates
    per-beat loggers in FactoryParams; preProcessors is invoked from
    reload paths; ToMap/extraArgs/StdFields are pure mapping helpers
    with no logger handle; getLogger() is the documented pre-
    SetLogger fallback; DefaultConfig runs before the beat-scoped
    logger is constructed). These match the dozens of other
    `logp.L()` callers across `heartbeat/` that pre-date the
    "accept *logp.Logger as a parameter" rule and a real refactor
    is tracked separately.

  * In `NewFactory`, hoisted `logger := logp.L()` to a local so the
    struct literal stays `goimports`-aligned with the `//nolint`
    directive sitting next to the assignment instead of inside the
    field block.

Verified locally with the same filter CI uses
(`--new-from-patch ... --new=false --whole-files`) against
golangci-lint v2.5.0 on `./heartbeat/...` and
`./x-pack/heartbeat/...`: 0 issues. All affected unit tests still
pass.

+++ heartbeat/monitors/wrappers/summarizer/summarizer.go
@@ -33,7 +33,7 @@ import (
 // this summary.
 type Summarizer struct {
 	rootJob        jobs.Job
-	contsRemaining uint16
+	contsRemaining int
 	mtx            *sync.Mutex
 	sf             stdfields.StdMonitorFields
 	mst            *monitorstate.Tracker
@@ -126,7 +126,7 @@ func (s *Summarizer) Wrap(j jobs.Job) jobs.Job {
 
 		s.contsRemaining-- // we just ran one cont, discount it
 		// these many still need to be processed
-		s.contsRemaining += uint16(len(conts))
+		s.contsRemaining += len(conts)
 
 		for _, plugin := range s.plugins {
 			actions := plugin.EachEvent(event, eventErr)
